// packages/prisma-oppulence/generated/zod/inputTypeSchemas/DealHistoryMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client'

import { z } from 'zod'
import { SortOrderSchema } from './SortOrderSchema'

export const DealHistoryMinOrderByAggregateInputSchema: z.ZodType<Prisma.DealHistoryMinOrderByAggregateInput> = z
    .object({
        id: z.lazy(() => SortOrderSchema).optional(),
        dealId: z.lazy(() => SortOrderSchema).optional(),
        field: z.lazy(() => SortOrderSchema).optional(),
        oldValue: z.lazy(() => SortOrderSchema).optional(),
        newValue: z.lazy(() => SortOrderSchema).optional(),
        description: z.lazy(() => SortOrderSchema).optional(),
        createdAt: z.lazy(() => SortOrderSchema).optional(),
        createdBy: z.lazy(() => SortOrderSchema).optional()
    })
    .strict()

export default DealHistoryMinOrderByAggregateInputSchema